Move Publication_Normalizer out of Datapoints.
