+++ includes/Modules/Reader_Revenue_Manager/Publication_Normalizer.php
@@ -1,14 +1,14 @@
 <?php
 /**
- * Class Google\Site_Kit\Modules\Reader_Revenue_Manager\Datapoints\Publication_Normalizer
+ * Class Google\Site_Kit\Modules\Reader_Revenue_Manager\Publication_Normalizer
  *
- * @package   Google\Site_Kit\Modules\Reader_Revenue_Manager\Datapoints
+ * @package   Google\Site_Kit\Modules\Reader_Revenue_Manager
  * @copyright 2026 Google LLC
  * @license   https://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://sitekit.withgoogle.com
  */
 
-namespace Google\Site_Kit\Modules\Reader_Revenue_Manager\Datapoints;
+namespace Google\Site_Kit\Modules\Reader_Revenue_Manager;
 
 use Google\Site_Kit_Dependencies\Google\Model;
 use Google\Site_Kit_Dependencies\Google\Service\SubscribewithGoogle\Publication;
